[benchmark] split up pkg + added benchmarks
