+++ benchmark/benchmark_test.go
@@ -2,11 +2,13 @@ package benchmark
 
 import (
 	"bytes"
+	"fmt"
+	"io"
 	"reflect"
+	"strings"
 	"testing"
 
-	"github.com/ShawnROGrady/benchplot/plot"
-	"github.com/ShawnROGrady/benchplot/plot/mock"
+	"golang.org/x/tools/benchmark/parse"
 )
 
 var parseBenchmarksTests = map[string]struct {
@@ -309,296 +311,174 @@ func TestGroupResults(t *testing.T) {
 	}
 }
 
-var splitGroupedResultsTests = map[string]struct {
-	groupedResults groupedResults
-	xName          string
-	yName          string
-	expectedSplit  map[string][]splitRes
-	expectErr      bool
-}{
-	"valid_x_valid_y": {
-		groupedResults: groupedResults{
-			"y=sin(x)": []benchRes{
-				{
-					inputs: benchInputs{
-						subs: []string{"areaUnder"},
-						varValues: []benchVarValue{
-							{name: "y", value: "sin(x)"},
-							{name: "delta", value: 0.001},
-						},
-					},
-					outputs: benchOutputs{N: 10, NsPerOp: 2000},
-				},
-				{
-					inputs: benchInputs{
-						subs: []string{"areaUnder"},
-						varValues: []benchVarValue{
-							{name: "y", value: "sin(x)"},
-							{name: "delta", value: 0.01},
-						},
-					},
-					outputs: benchOutputs{N: 100, NsPerOp: 200},
-				},
-			},
-			"y=2x+3": []benchRes{
-				{
-					inputs: benchInputs{
-						subs: []string{"areaUnder"},
-						varValues: []benchVarValue{
-							{name: "y", value: "2x+3"},
-							{name: "delta", value: 0.001},
-						},
-					},
-					outputs: benchOutputs{N: 5, NsPerOp: 1000},
-				},
-				{
-					inputs: benchInputs{
-						subs: []string{"areaUnder"},
-						varValues: []benchVarValue{
-							{name: "y", value: "2x+3"},
-							{name: "delta", value: 0.01},
-						},
-					},
-					outputs: benchOutputs{N: 10, NsPerOp: 100},
-				},
-			},
-		},
-		xName: "delta", yName: TimeName,
-		expectedSplit: map[string][]splitRes{
-			"y=sin(x)": []splitRes{{x: 0.001, y: float64(2000)}, {x: 0.01, y: float64(200)}},
-			"y=2x+3":   []splitRes{{x: 0.001, y: float64(1000)}, {x: 0.01, y: float64(100)}},
-		},
-	},
-	"invalid_x": {
-		groupedResults: groupedResults{
-			"y=sin(x)": []benchRes{
-				{
-					inputs: benchInputs{
-						subs: []string{"areaUnder"},
-						varValues: []benchVarValue{
-							{name: "y", value: "sin(x)"},
-							{name: "delta", value: 0.001},
-						},
-					},
-					outputs: benchOutputs{N: 10, NsPerOp: 2000},
-				},
-			},
-		},
-		xName: "invalid_var", yName: TimeName,
-		expectErr: true,
-	},
-	"invalid_y": {
-		groupedResults: groupedResults{
-			"y=sin(x)": []benchRes{
-				{
-					inputs: benchInputs{
-						subs: []string{"areaUnder"},
-						varValues: []benchVarValue{
-							{name: "y", value: "sin(x)"},
-							{name: "delta", value: 0.001},
-						},
-					},
-					outputs: benchOutputs{N: 10, NsPerOp: 2000},
-				},
-			},
-		},
-		xName: "delta", yName: "invalid_var",
-		expectErr: true,
-	},
-}
+var parseBenchmarksErr error
 
-func TestSplitGroupedResults(t *testing.T) {
-	for testName, testCase := range splitGroupedResultsTests {
-		t.Run(testName, func(t *testing.T) {
-			splitGrouped, err := testCase.groupedResults.splitTo(testCase.xName, testCase.yName)
-			if err != nil {
-				if !testCase.expectErr {
-					t.Errorf("unexpected error: %s", err)
-				}
-				return
-			}
+func BenchmarkParseBenchmarks(b *testing.B) {
+	var (
+		allNumBenchmarks     = []int{1, 2, 3, 4, 5}
+		allCasesPerBenchmark = []int{5, 10, 15, 20, 25}
+	)
 
-			if testCase.expectErr {
-				t.Fatalf("unexpectedly no error")
-			}
-
-			if !reflect.DeepEqual(splitGrouped, testCase.expectedSplit) {
-				t.Errorf("unexpected split grouped results\nexpected:\n%#v\nactual:\n%#v", testCase.expectedSplit, splitGrouped)
+	for _, numBenchmarks := range allNumBenchmarks {
+		b.Run(fmt.Sprintf("num_benchmarks=%d", numBenchmarks), func(b *testing.B) {
+			for _, casesPerBench := range allCasesPerBenchmark {
+				b.Run(fmt.Sprintf("cases_per_bench=%d", casesPerBench), func(b *testing.B) {
+					benchmarkParseBenchmarks(b, numBenchmarks, casesPerBench)
+				})
 			}
 		})
 	}
 }
 
-var sampleBenchmark = Benchmark{
-	Name: "BenchmarkMath",
-	results: []benchRes{
-		{
-			inputs: benchInputs{
-				subs: []string{"areaUnder"},
-				varValues: []benchVarValue{
-					{name: "y", value: "sin(x)"},
-					{name: "delta", value: 0.001},
-					{name: "start_x", value: -2},
-					{name: "end_x", value: 1},
-				},
-			},
-			outputs: benchOutputs{N: 10, NsPerOp: 2000},
-		},
-		{
-			inputs: benchInputs{
-				subs: []string{"areaUnder"},
-				varValues: []benchVarValue{
-					{name: "y", value: "sin(x)"},
-					{name: "delta", value: 0.01},
-					{name: "start_x", value: -2},
-					{name: "end_x", value: 1},
-				},
-			},
-			outputs: benchOutputs{N: 100, NsPerOp: 200},
-		},
-		{
-			inputs: benchInputs{
-				subs: []string{"areaUnder"},
-				varValues: []benchVarValue{
-					{name: "y", value: "2x+3"},
-					{name: "delta", value: 0.001},
-					{name: "start_x", value: -2},
-					{name: "end_x", value: 1},
-				},
-			},
-			outputs: benchOutputs{N: 5, NsPerOp: 1000},
-		},
-		{
-			inputs: benchInputs{
-				subs: []string{"areaUnder"},
-				varValues: []benchVarValue{
-					{name: "y", value: "2x+3"},
-					{name: "delta", value: 0.01},
-					{name: "start_x", value: -2},
-					{name: "end_x", value: 1},
-				},
-			},
-			outputs: benchOutputs{N: 10, NsPerOp: 100},
-		},
-	},
-}
-
-var plotScatterTests = map[string]struct {
-	benchmark      Benchmark
-	groupBy        []string
-	xName          string
-	yName          string
-	expectedData   map[string]plot.NumericData
-	expectedTitle  string
-	expectedXLabel string
-	expectedYLabel string
-	expectErr      bool
-}{
-	"x=float64,y=float64": {
-		benchmark: sampleBenchmark,
-		groupBy:   []string{"y"},
-		xName:     "delta", yName: TimeName,
-		expectedData: map[string]plot.NumericData{
-			"y=sin(x)": plot.NumericData{
-				X: []float64{0.001, 0.01},
-				Y: []float64{2000, 200},
-			},
-			"y=2x+3": plot.NumericData{
-				X: []float64{0.001, 0.01},
-				Y: []float64{1000, 100},
-			},
-		},
-		expectedTitle:  "BenchmarkMath",
-		expectedXLabel: "delta",
-		expectedYLabel: TimeName,
-	},
-	"x=int,y=int": {
-		benchmark: sampleBenchmark,
-		groupBy:   []string{"y"},
-		xName:     "start_x", yName: RunsName,
-		expectedData: map[string]plot.NumericData{
-			"y=sin(x)": plot.NumericData{
-				X: []float64{-2, -2},
-				Y: []float64{10, 100},
-			},
-			"y=2x+3": plot.NumericData{
-				X: []float64{-2, -2},
-				Y: []float64{5, 10},
-			},
-		},
-		expectedTitle:  "BenchmarkMath",
-		expectedXLabel: "start_x",
-		expectedYLabel: RunsName,
-	},
-	"x=float64,y=uint64": {
-		benchmark: sampleBenchmark,
-		groupBy:   []string{"y"},
-		xName:     "delta", yName: NumAllocsName,
-		expectedData: map[string]plot.NumericData{
-			"y=sin(x)": plot.NumericData{
-				X: []float64{0.001, 0.01},
-				Y: []float64{0, 0},
-			},
-			"y=2x+3": plot.NumericData{
-				X: []float64{0.001, 0.01},
-				Y: []float64{0, 0},
-			},
-		},
-		expectedTitle:  "BenchmarkMath",
-		expectedXLabel: "delta",
-		expectedYLabel: NumAllocsName,
-	},
-	"x=string,y=float64": {
-		benchmark: sampleBenchmark,
-		groupBy:   []string{"start_x"},
-		xName:     "y", yName: TimeName,
-		expectErr: true,
-	},
-	"invalid_x_name": {
-		benchmark: sampleBenchmark,
-		groupBy:   []string{"start_x"},
-		xName:     "invalid_name", yName: TimeName,
-		expectErr: true,
-	},
-	"invalid_y_name": {
-		benchmark: sampleBenchmark,
-		groupBy:   []string{"start_x"},
-		xName:     "delta", yName: "invalid_name",
-		expectErr: true,
-	},
-}
-
-func TestPlotScatter(t *testing.T) {
-	for testName, testCase := range plotScatterTests {
-		t.Run(testName, func(t *testing.T) {
-			p := &mock.Plotter{
-				PlotScatterFn: func(data map[string]plot.NumericData, title string, xLabel string, yLabel string) error {
-					// validate args
-					if !reflect.DeepEqual(data, testCase.expectedData) {
-						t.Errorf("unexpected plot data\nexpected:\n%v\nactual:\n%v", testCase.expectedData, data)
-					}
-					if title != testCase.expectedTitle {
-						t.Errorf("unexpected title\nexpected:\n%s\nactual:\n%s", testCase.expectedTitle, title)
-					}
-					if xLabel != testCase.expectedXLabel {
-						t.Errorf("unexpected xLabel\nexpected:\n%s\nactual:\n%s", testCase.expectedXLabel, xLabel)
-					}
-					if yLabel != testCase.expectedYLabel {
-						t.Errorf("unexpected yLabel\nexpected:\n%s\nactual:\n%s", testCase.expectedYLabel, yLabel)
-					}
-					return nil
-				},
-			}
-
-			err := testCase.benchmark.PlotScatter(p, testCase.groupBy, testCase.xName, testCase.yName)
-			if err != nil {
-				if !testCase.expectErr {
-					t.Errorf("unexpected error: %s", err)
+func benchmarkParseBenchmarks(b *testing.B, numBenchmarks, casesPerBench int) {
+	b.Helper()
+	newReader := func() io.Reader {
+		var buf bytes.Buffer
+		for i := 0; i < numBenchmarks; i++ {
+			benchName := fmt.Sprintf("BenchmarkMethod%d", i)
+			for j := 0; j < casesPerBench; j++ {
+				bench := &parse.Benchmark{
+					Name:    fmt.Sprintf("%s/var1=%d/var2=%d", benchName, j, j),
+					N:       j,
+					NsPerOp: float64(j),
 				}
-				return
+				if _, err := buf.WriteString(fmt.Sprintf("%s\n", bench)); err != nil {
+					b.Fatalf("err constructing input: %s", err)
+				}
 			}
-			if testCase.expectErr {
-				t.Error("unexpectedly no error")
+		}
+		return &buf
+	}
+
+	var err error
+	var benches []Benchmark
+	for i := 0; i < b.N; i++ {
+		b.StopTimer()
+		r := newReader()
+		b.StartTimer()
+
+		benches, err = ParseBenchmarks(r)
+		if err != nil {
+			b.Fatalf("unexpected error: %s", err)
+		}
+		if len(benches) != numBenchmarks {
+			b.Fatalf("unexpected number of benchmarks (expected=%d, actual=%d)", numBenchmarks, len(benches))
+		}
+	}
+	parseBenchmarksErr = err
+}
+
+var benchGroupResultsRes groupedResults
+
+func BenchmarkGroupResults(b *testing.B) {
+	var (
+		allNumValues    = []int{1, 2, 3, 4, 5, 10, 20}
+		allGroupByCount = []int{1, 2, 3}
+		allNumResults   = []int{5, 10, 15, 20, 25}
+	)
+
+	for _, numResults := range allNumResults {
+		b.Run(fmt.Sprintf("num_results=%d", numResults), func(b *testing.B) {
+			for _, numValues := range allNumValues {
+				b.Run(fmt.Sprintf("num_values=%d", numValues), func(b *testing.B) {
+					for _, groupByCount := range allGroupByCount {
+						b.Run(fmt.Sprintf("group_by_count=%d", groupByCount), func(b *testing.B) {
+							if groupByCount > numValues {
+								b.Skip("skipping due to groupByCount>numValues")
+							}
+							benchmarkGroupResults(b, numResults, numValues, groupByCount)
+						})
+					}
+				})
+			}
+		})
+	}
+}
+
+func benchmarkGroupResults(b *testing.B, numResults, numValues, groupByCount int) {
+	b.Helper()
+	var (
+		bench   = newTestBenchmark(numResults, numValues)
+		groupBy = make([]string, groupByCount)
+	)
+	for i := 0; i < groupByCount; i++ {
+		groupBy[i] = fmt.Sprintf("var%d", i)
+	}
+
+	var groupRes groupedResults
+	for i := 0; i < b.N; i++ {
+		groupRes = bench.groupResults(groupBy)
+	}
+	benchGroupResultsRes = groupRes
+}
+
+func newTestBenchmark(numResults, numValues int) Benchmark {
+	res := make([]benchRes, numResults)
+	for i := 0; i < numResults; i++ {
+		var (
+			outputs   = benchOutputs{N: i, NsPerOp: float64(i)}
+			varValues = make([]benchVarValue, numValues)
+		)
+		for j := 0; j < numValues; j++ {
+			varValues[j] = benchVarValue{
+				name:  fmt.Sprintf("var%d", j),
+				value: j,
+			}
+		}
+		res[i] = benchRes{
+			inputs: benchInputs{
+				varValues: varValues,
+			},
+			outputs: outputs,
+		}
+	}
+	return Benchmark{
+		Name:    "BenchmarkSomeMethod",
+		results: res,
+	}
+}
+
+var parseInfoErr error
+
+func BenchmarkParseInfo(b *testing.B) {
+	var (
+		dTypes = map[string]func(varName string) string{
+			"int": func(varName string) string {
+				return fmt.Sprintf("%s=%d", varName, 1)
+			},
+			"float64": func(varName string) string {
+				return fmt.Sprintf("%s=%f", varName, 1.1)
+			},
+			"bool": func(varName string) string {
+				return fmt.Sprintf("%s=%t", varName, true)
+			},
+			"string": func(varName string) string {
+				return fmt.Sprintf("%s=%s", varName, "foo")
+			},
+		}
+		allNumValues = []int{1, 2, 3, 4, 5, 10, 20}
+	)
+
+	for _, numValues := range allNumValues {
+		b.Run(fmt.Sprintf("num_values=%d", numValues), func(b *testing.B) {
+			for dtype, fn := range dTypes {
+				b.Run(fmt.Sprintf("dtype=%s", dtype), func(b *testing.B) {
+					s := make([]string, numValues+1)
+					s[0] = "BenchmarkSomeMethod"
+					for i := 1; i <= numValues; i++ {
+						s[i] = fn(fmt.Sprintf("var%d", i))
+					}
+					input := strings.Join(s, "/")
+
+					var err error
+					for i := 0; i < b.N; i++ {
+						_, _, err = parseInfo(input)
+						if err != nil {
+							b.Fatalf("unexpected error: %s", err)
+						}
+					}
+					parseInfoErr = err
+				})
 			}
 		})
 	}
